slight refactoring
  - config in own package
  - protobuf in own package
